Fix typos and move around types
