+++ web/packages/teleport/src/services/auth/auth.ts
@@ -16,11 +16,7 @@ limitations under the License.
 
 import api from 'teleport/services/api';
 import cfg from 'teleport/config';
-import {
-  DeviceType,
-  DeviceUsage,
-  IsMfaRequiredRequest,
-} from 'teleport/services/mfa';
+import { DeviceType, DeviceUsage } from 'teleport/services/mfa';
 
 import makePasswordToken from './makePasswordToken';
 import { makeChangedUserAuthn } from './make';
@@ -247,3 +243,47 @@ function base64EncodeUnicode(str: string) {
 }
 
 export default auth;
+
+export type IsMfaRequiredRequest =
+  | IsMfaRequiredDatabase
+  | IsMfaRequiredNode
+  | IsMfaRequiredKube
+  | IsMfaRequiredWindowsDesktop;
+
+export type IsMfaRequiredDatabase = {
+  database: {
+    // service_name is the database service name.
+    service_name: string;
+    // protocol is the type of the database protocol.
+    protocol: string;
+    // username is an optional database username.
+    username?: string;
+    // database_name is an optional database name.
+    database_name?: string;
+  };
+};
+
+export type IsMfaRequiredNode = {
+  node: {
+    // node_name can be node's hostname or UUID.
+    node_name: string;
+    // login is the OS login name.
+    login: string;
+  };
+};
+
+export type IsMfaRequiredWindowsDesktop = {
+  windows_desktop: {
+    // desktop_name is the Windows Desktop server name.
+    desktop_name: string;
+    // login is the Windows desktop user login.
+    login: string;
+  };
+};
+
+export type IsMfaRequiredKube = {
+  kube: {
+    // cluster_name is the name of the kube cluster.
+    cluster_name: string;
+  };
+};
