+++ web/packages/teleport/src/services/mfa/types.ts
@@ -45,47 +45,3 @@ export type DeviceUsage = 'passwordless' | 'mfa';
 export type MfaAuthnResponse =
   | { totp_code: string }
   | { webauthn_response: WebauthnAssertionResponse };
-
-export type IsMfaRequiredDatabase = {
-  database: {
-    // service_name is the database service name.
-    service_name: string;
-    // protocol is the type of the database protocol.
-    protocol: string;
-    // username is an optional database username.
-    username?: string;
-    // database_name is an optional database name.
-    database_name?: string;
-  };
-};
-
-export type IsMfaRequiredNode = {
-  node: {
-    // node_name can be node's hostname or UUID.
-    node_name: string;
-    // login is the OS login name.
-    login: string;
-  };
-};
-
-export type IsMfaRequiredWindowsDesktop = {
-  windows_desktop: {
-    // desktop_name is the Windows Desktop server name.
-    desktop_name: string;
-    // login is the Windows desktop user login.
-    login: string;
-  };
-};
-
-export type IsMfaRequiredKube = {
-  kube: {
-    // cluster_name is the name of the kube cluster.
-    cluster_name: string;
-  };
-};
-
-export type IsMfaRequiredRequest =
-  | IsMfaRequiredDatabase
-  | IsMfaRequiredNode
-  | IsMfaRequiredKube
-  | IsMfaRequiredWindowsDesktop;
